+++ templates/email/tracking-info.php
@@ -4,8 +4,8 @@
  *
  * Shows tracking information in the HTML order email
  *
- * @author  Remi Corson
- * @version 1.0
+ * @author  Nicolas Mollet
+ * @version 1.0.1
  */
 
 if ( ! defined( 'ABSPATH' ) ) {
@@ -13,39 +13,18 @@ if ( ! defined( 'ABSPATH' ) ) {
 }
 
 if ( $tracking_items ) : ?>
-	<h2><?php echo apply_filters( 'woocommerce_la_poste_tracking_my_orders_title', __( 'Tracking Information', 'tmsm-woocommerce-laposte-tracking' ) ); ?></h2>
+	<h2><?php echo apply_filters( 'woocommerce_la_poste_tracking_my_orders_title',
+			__( 'Tracking Information', 'tracking-la-poste-for-woocommerce' ) ); ?></h2>
+	<?php
+	foreach ( $tracking_items as $tracking_item ) {
 
-	<table class="td" cellspacing="0" cellpadding="6" style="width: 100%;" border="1">
+		?>
+		<p><?php printf(__( 'Your product has been shipped. The tracking number is: %s', 'tmsm-woocommerce-laposte-tracking' ), '<a href="'.esc_url( $tracking_item[ 'formatted_tracking_link' ] ).'">'.esc_html( $tracking_item['tracking_number'] ).'</a>'); ?>
+		</p>
 
-		<thead>
-			<tr>
-				<th class="tracking-number" scope="col" class="td" style="text-align: left; font-family: 'Helvetica Neue', Helvetica, Roboto, Arial, sans-serif; color: #737373; border: 1px solid #e4e4e4; padding: 12px;"><?php _e( 'Tracking Number', 'tmsm-woocommerce-laposte-tracking' ); ?></th>
-				<th class="date-shipped" scope="col" class="td" style="text-align: left; font-family: 'Helvetica Neue', Helvetica, Roboto, Arial, sans-serif; color: #737373; border: 1px solid #e4e4e4; padding: 12px;"><?php _e( 'Date', 'tmsm-woocommerce-laposte-tracking' ); ?></th>
-				<th class="tracking-status" scope="col" class="td" style="text-align: left; font-family: 'Helvetica Neue', Helvetica, Roboto, Arial, sans-serif; color: #737373; border: 1px solid #e4e4e4; padding: 12px;"><?php _e( 'Status', 'tmsm-woocommerce-laposte-tracking' ); ?></th>
-				<th class="order-actions" scope="col" class="td" style="text-align: left; font-family: 'Helvetica Neue', Helvetica, Roboto, Arial, sans-serif; color: #737373; border: 1px solid #e4e4e4; padding: 12px;">&nbsp;</th>
-			</tr>
-		</thead>
-
-		<tbody><?php
-			foreach ( $tracking_items as $tracking_item ) {
-				
-				?><tr class="tracking">
-					<td class="tracking-number" data-title="<?php _e( 'Tracking Number', 'tmsm-woocommerce-laposte-tracking' ); ?>" style="text-align: left; font-family: 'Helvetica Neue', Helvetica, Roboto, Arial, sans-serif; color: #737373; border: 1px solid #e4e4e4; padding: 12px;">
-						<?php echo esc_html( $tracking_item[ 'tracking_number' ] ); ?>
-					</td>
-					<td class="date-shipped" data-title="<?php _e( 'Date', 'tmsm-woocommerce-laposte-tracking' ); ?>" style="text-align: left; font-family: 'Helvetica Neue', Helvetica, Roboto, Arial, sans-serif; color: #737373; border: 1px solid #e4e4e4; padding: 12px;">
-						<time datetime="<?php echo date( 'Y-m-d', $tracking_item[ 'date_shipped' ] ); ?>" title="<?php echo date( 'Y-m-d', $tracking_item[ 'date_shipped' ] ); ?>"><?php echo date_i18n( get_option( 'date_format' ), $tracking_item[ 'date_shipped' ] ); ?></time>
-					</td>
-					<td class="tracking-status" data-title="<?php _e( 'Status', 'tmsm-woocommerce-laposte-tracking' ); ?>" style="text-align: left; font-family: 'Helvetica Neue', Helvetica, Roboto, Arial, sans-serif; color: #737373; border: 1px solid #e4e4e4; padding: 12px;">
-						<?php echo esc_html( $tracking_item[ 'tracking_message' ] ); ?> (<?php echo esc_html( $tracking_item[ 'tracking_status' ] ); ?>)
-					</td>
-					<td class="order-actions" style="text-align: center; font-family: 'Helvetica Neue', Helvetica, Roboto, Arial, sans-serif; color: #737373; border: 1px solid #e4e4e4; padding: 12px;">
-							<?php if( $tracking_item[ 'formatted_tracking_link' ] ) : ?><a href="<?php echo esc_url( $tracking_item[ 'formatted_tracking_link' ] ); ?>" target="_blank"><?php _e( 'Track', 'tmsm-woocommerce-laposte-tracking' ); ?></a><?php else: ?><em><?php _e( 'Tracking unavailable', 'tmsm-woocommerce-laposte-tracking' ); ?></em><?php endif; ?>
-					</td>
-				</tr><?php
-			}
-		?></tbody>
-	</table>
+		<?php
+	}
+	?>
 
 <?php
 endif;